requirements.txt
azure-functions
pymongo
bcrypt
python-jose
sendgrid
python-dotenv
